Added values for player data
Added values for player data (and structure to replace the hardcoded record with database connection) in PlayPage.cs. Next commit will be after all play page switching has been coded.

+++ testing_v2/PatientData.cs
@@ -37,7 +37,7 @@ namespace testing_v2
         public List<String> OxygenData { get; set; }
 
         // 16 values in bloodwork
-        public List<float> BloodworkData { get; set; }
+        public List<double> BloodworkData { get; set; }
 
         // XRayImage
         public Texture2D XRayImage;
